requirements.txt
Flask
ipfshttpclient
cryptography